Separate out client capabilities test into its own file
