+++ packages/core/tests/integration/elicitation-e2e.test.ts
@@ -11,7 +11,8 @@ import {
 } from "../../src/index.js";
 
 describe("Elicitation E2E Tests", () => {
-  test("E2E: ctx.elicit() throws when client has elicitation capability", async () => {
+  // FIXME - this does not actually test ctx.elicit
+  test("E2E: ctx.elicit() does not throw when client has elicitation capability", async () => {
     const server = new McpServer({
       name: "elicitation-test-server",
       version: "1.0.0",
@@ -490,103 +491,6 @@ describe("Elicitation E2E Tests", () => {
     expect(toolResponse.status).toBe(200);
     const toolResult = await toolResponse.json();
     expect(toolResult.result.content[0].text).toBe("Hello, Alice!");
-  });
-
-  test("E2E: client capabilities are properly stored and retrieved", async () => {
-    const server = new McpServer({
-      name: "elicitation-test-server",
-      version: "1.0.0",
-      schemaAdapter: (s) => z.toJSONSchema(s as z.ZodType),
-    });
-
-    const clientRequestAdapter = new InMemoryClientRequestAdapter();
-    const sessionAdapter = new InMemorySessionAdapter({
-      maxEventBufferSize: 1024,
-    });
-
-    const transport = new StreamableHttpTransport({
-      clientRequestAdapter,
-      sessionAdapter,
-    });
-
-    const handler = transport.bind(server);
-
-    server.tool("test-capabilities", {
-      description: "Test capability detection",
-      inputSchema: z.object({}),
-      handler: async (_, ctx) => {
-        const hasElicitation = ctx.client.supports("elicitation");
-        const hasRoots = ctx.client.supports("roots");
-        const hasSampling = ctx.client.supports("sampling");
-        const hasUnknown = ctx.client.supports("unknown");
-
-        return {
-          content: [
-            {
-              type: "text",
-              text: `elicitation:${hasElicitation},roots:${hasRoots},sampling:${hasSampling},unknown:${hasUnknown}`,
-            },
-          ],
-        };
-      },
-    });
-
-    // Test 1: Initialize with multiple capabilities
-    const initializeRequest = {
-      jsonrpc: "2.0",
-      id: 1,
-      method: "initialize",
-      params: {
-        clientInfo: { name: "test-client", version: "1.0.0" },
-        protocolVersion: "2025-06-18",
-        capabilities: {
-          elicitation: {},
-          roots: {},
-          // Note: no sampling capability
-        },
-      },
-    };
-
-    const initResponse = await handler(
-      new Request("http://localhost:3000/", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(initializeRequest),
-      }),
-    );
-
-    expect(initResponse.status).toBe(200);
-    const sessionId = initResponse.headers.get("mcp-session-id")!;
-
-    // Test 2: Call tool to check capabilities
-    const toolResponse = await handler(
-      new Request("http://localhost:3000/", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          "mcp-session-id": sessionId,
-        },
-        body: JSON.stringify({
-          jsonrpc: "2.0",
-          id: "tool-call-1",
-          method: "tools/call",
-          params: {
-            name: "test-capabilities",
-            arguments: {},
-          },
-        }),
-      }),
-    );
-
-    expect(toolResponse.status).toBe(200);
-    const toolResult = await toolResponse.json();
-
-    // Verify that capabilities are correctly detected
-    expect(toolResult.result.content[0].text).toBe(
-      "elicitation:true,roots:true,sampling:false,unknown:false",
-    );
   });
 
   test("E2E: full elicitation flow with client decline", async () => {
